Refactor common files
