+++ bodo/tests/test_io.py
@@ -1,9 +1,7 @@
 """Tests I/O for CSV, NP IO, etc."""
 
-import gzip
 import io
 import os
-import shutil
 import subprocess
 import unittest
 
@@ -22,12 +20,14 @@ from bodo.tests.utils import (
     _get_dist_arg,
     _test_equal_guard,
     check_func,
+    compress_dir,
     count_array_REPs,
     count_parfor_REPs,
     get_rank,
     get_start_end,
     pytest_mark_one_rank,
     reduce_sum,
+    uncompress_dir,
 )
 from bodo.utils.testing import ensure_clean
 from bodo.utils.typing import BodoError
@@ -55,33 +55,6 @@ def remove_files(file_names):
     if bodo.get_rank() == 0:
         for fname in file_names:
             os.remove(fname)
-    bodo.barrier()
-
-
-def compress_dir(dir_name):
-    if bodo.get_rank() == 0:
-        for fname in [
-            f
-            for f in os.listdir(dir_name)
-            if f.endswith(".csv") and os.path.getsize(os.path.join(dir_name, f)) > 0
-        ]:
-            full_fname = os.path.join(dir_name, fname)
-            out_fname = full_fname + ".gz"
-            with open(full_fname, "rb") as f_in:
-                with gzip.open(out_fname, "wb") as f_out:
-                    shutil.copyfileobj(f_in, f_out)
-            os.remove(full_fname)
-    bodo.barrier()
-
-
-def uncompress_dir(dir_name):
-    if bodo.get_rank() == 0:
-        for fname in [f for f in os.listdir(dir_name) if f.endswith(".gz")]:
-            full_fname = os.path.join(dir_name, fname)
-            with gzip.open(full_fname, "rb") as f_in:
-                with open(full_fname.removesuffix(".gz"), "wb") as f_out:
-                    shutil.copyfileobj(f_in, f_out)
-            os.remove(full_fname)
     bodo.barrier()
